Lift fn_801C2914 97.18->98.12: drop (s32) intermediates on obj+2/obj+4 f32->s16 stores (extsh x2, recipe #83c)

+++ src/main/dll/DF/DFlantern.c
@@ -246,12 +246,12 @@ void fn_801C2914(int obj)
   trigA = fn_80293E80((lbl_803E4E60 * (f32)state->orbitB) / lbl_803E4E64);
   trigB = fn_80293E80((lbl_803E4E60 * (f32)state->orbitA) / lbl_803E4E64);
   trigB = trigB + trigA;
-  *(s16 *)(obj + 4) = (s32)(lbl_803E4E68 * trigB);
+  *(s16 *)(obj + 4) = lbl_803E4E68 * trigB;
 
   trigA = fn_80293E80((lbl_803E4E60 * (f32)state->orbitC) / lbl_803E4E64);
   trigB = fn_80293E80((lbl_803E4E60 * (f32)state->orbitA) / lbl_803E4E64);
   trigB = trigB + trigA;
-  *(s16 *)(obj + 2) = (s32)(lbl_803E4E68 * trigB);
+  *(s16 *)(obj + 2) = lbl_803E4E68 * trigB;
 
   ObjAnim_AdvanceCurrentMove(lbl_803E4E6C,timeDelta,obj,(ObjAnimEventList *)animEvents);
   if (player != NULL) {
